feat(cafm): all four service pages on the shared scaffold; v2.23.0+39
Cleaning, landscaping and facade now use the same scaffold security got: an
identity band whose every number opens the records it counts, a record-kind
rail, per-kind search, and "المزيد" paging on the long lists — while keeping
each service's own logic (the cleaning audit accept/dispute sheet, the plant
health colours, the species catalogue cards, the facade wind-safety pills).

All four are consistent now, and none of them dumps an unbounded list.

+++ lib/screens/client_activity_screen.dart
@@ -116,7 +116,7 @@ class _ClientActivityScreenState extends State<ClientActivityScreen> {
   /// The "right now" band — what a client checks before anything else.
   Widget _pulse(Map st) {
     final byHour = (st['by_hour'] as Map?) ?? const {};
-    final maxH = byHour.values.fold<int>(1, (a, v) => (v as int) > a ? v as int : a);
+    final maxH = byHour.values.fold<int>(1, (a, v) => (v as int) > a ? v : a);
     return Container(
       padding: const EdgeInsets.fromLTRB(14, 12, 14, 12),
       decoration: BoxDecoration(
